% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_inputs.R
\name{create_calsim_non_cvpia_nodes_tidy}
\alias{create_calsim_non_cvpia_nodes_tidy}
\title{Create CalSim non-CVPIA nodes tidy data}
\usage{
create_calsim_non_cvpia_nodes_tidy()
}
\value{
A tidy data frame with columns: year, location, value, scenario, run, and performance_metric.
}
\description{
This function retrieves and processes data related to hydropower generation, water deliveries for agriculture, municipal and industrial (M&I) use,
and refuge water supply and delivery from CalSim non-CVPIA nodes. The function performs data transformation and returns a tidy data frame
containing the processed data for further analysis.
}
\examples{
create_calsim_non_cvpia_nodes_tidy()

}
