% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_inputs.R
\name{create_model_inputs_tidy_df}
\alias{create_model_inputs_tidy_df}
\title{Create Model Inputs Tidy Data Frame}
\usage{
create_model_inputs_tidy_df(model_parameters, scenario_name, selected_run)
}
\arguments{
\item{model_parameters}{A list containing the model parameters, including spawning habitat, in-channel habitat for fry, in-channel habitat for juvenile, floodplain habitat, and other parameters.}

\item{scenario_name}{A character string specifying the name of the scenario.}

\item{selected_run}{A character string specifying the selected run.}
}
\value{
A tidy data frame containing the performance metrics.
}
\description{
This function generates a tidy data frame containing various performance metrics related to habitat and flow for a given set of model parameters, scenario name, and selected run.
}
\examples{
model_params <- fallRunDSM::r_to_r_baseline_params
scenario <- "Baseline"
run <- "fall"
create_model_inputs_tidy_df(model_params, scenario, run)

}
