% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_results.R
\name{create_model_results_dataframe}
\alias{create_model_results_dataframe}
\title{Create model results dataframe}
\usage{
create_model_results_dataframe(model_results, scenario_name, chinook_run)
}
\arguments{
\item{model_results}{A model output from the Recovery life cycle models (fall, spring, winter or late fall)}

\item{scenario_name}{The name of the scenario that produced the model results.}
}
\value{
A dataframe containing the following columns(location, year, scenario, performance_metric, value)
}
\description{
Create model results dataframe
}
\examples{
baseline_seeds <- fallRunDSM::fall_run_model(mode = "seed", ..params = fallRunDSM::r_to_r_baseline_params)
baseline_model_results <- fallRunDSM::fall_run_model(mode = "simulate", ..params = fallRunDSM::r_to_r_baseline_params,
                                                     seeds = baseline_seeds)
process_model_results(baseline_model_results, "Baseline")
}
