% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_biological_pms.R
\name{produce_categorical_return_age_pm}
\alias{produce_categorical_return_age_pm}
\title{This function calculates the number of locations meeting the categorical return age criteria based on model results.}
\usage{
produce_categorical_return_age_pm(model_results_df)
}
\arguments{
\item{model_results_df}{A data frame containing model results.}
}
\value{
A data frame with the count of locations meeting the categorical return age criteria per scenario and run.
}
\description{
This function calculates the number of locations meeting the categorical return age criteria based on model results.
}
\examples{
produce_categorical_return_age_pm(model_results)

}
