% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_biological_pms.R
\name{produce_crr_pm}
\alias{produce_crr_pm}
\title{Calculate CRR (Total Adult to Returning Natural Adult)}
\usage{
produce_crr_pm(model_results_df)
}
\arguments{
\item{model_results_df}{A data frame containing model results.}
}
\value{
A data frame with the calculated average, minimum, and maximum annual CRR per scenario.
}
\description{
This function calculates the average, minimum, and maximum annual CRR (Total Adult to Returning Natural Adult) based on model results.
}
\examples{
produce_crr_pm(model_results)

}
