% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_biological_pms.R
\name{produce_floodplain_over_inchannel_habitat}
\alias{produce_floodplain_over_inchannel_habitat}
\title{Calculate Carrying Capacity vs. Abundance}
\usage{
produce_floodplain_over_inchannel_habitat(
  model_results_df,
  model_parameters,
  selected_run
)
}
\arguments{
\item{model_results_df}{The model results dataframe.}

\item{model_parameters}{The model parameters.}

\item{run}{The selected run.}
}
\value{
A tibble with carrying capacity vs. abundance values.
}
\description{
This function calculates the carrying capacity vs. abundance based on the provided model
results dataframe, model parameters, and run. Only spawning habitat is considered, not holding.
}
\examples{
model_results <- data.frame(
  spawning_habitat = c(100, 200, 150, 250),
  location = c("Location A", "Location B", "Location A", "Location B"),
  year_date = c("Jan-2021", "Jan-2021", "Feb-2021", "Feb-2021")
)
model_parameters <- list(spawning_habitat = data.frame(location = c("Location A", "Location B"), year_date = "Jan-2021", value = 300))

produce_carrying_capacity_vs_abundance(model_results, model_parameters, "fall")
}
