% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_biological_pms.R
\name{produce_growth_rate_pm}
\alias{produce_growth_rate_pm}
\title{Calculate Growth Rates of Natural Spawners}
\usage{
produce_growth_rate_pm(model_results_df)
}
\arguments{
\item{model_results_df}{A data frame containing model results.}
}
\value{
A data frame with the calculated average, minimum, and maximum annual growth rates per scenario.
}
\description{
This function calculates the average, minimum, and maximum annual growth rates of natural spawners based on model results.
}
\examples{
produce_growth_rate_pm(model_results)

}
