% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_biological_pms.R
\name{produce_independent_pops_per_diversity_group_pm}
\alias{produce_independent_pops_per_diversity_group_pm}
\title{Calculate Number of Independent Populations per Diversity Group}
\usage{
produce_independent_pops_per_diversity_group_pm(model_results_df, selected_run)
}
\arguments{
\item{model_results_df}{A data frame containing model results.}

\item{selected_run}{A character string specifying the selected run ("fall", "spring", or "winter").}
}
\value{
A data frame with the calculated number of independent populations per diversity group.
}
\description{
This function calculates the number of independent populations per diversity group based on model results and a selected run.
}
\examples{
produce_independent_pops_per_diversity_group_pm(model_results, "fall")

}
