% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_biological_pms.R
\name{produce_independent_pops_pm}
\alias{produce_independent_pops_pm}
\title{Calculate Number of Independent Populations}
\usage{
produce_independent_pops_pm(model_results_df)
}
\arguments{
\item{model_results_df}{A data frame containing model results.}
}
\value{
A data frame with the calculated number of independent populations per scenario.
}
\description{
This function calculates the number of independent populations based on model results.
}
\examples{
produce_independent_pops_pm(model_results)

}
