% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_biological_pms.R
\name{produce_marine_nutrient_pm}
\alias{produce_marine_nutrient_pm}
\title{Calculate marine nutrient production per square meter}
\usage{
produce_marine_nutrient_pm(model_results_df, scenario_name)
}
\arguments{
\item{model_results_df}{A data frame containing model results.}

\item{scenario_name}{A character string specifying the scenario name.}
}
\value{
A modified data frame with additional columns representing marine nutrient production per square meter.
}
\description{
This function calculates the marine nutrient production per square meter based on model results and scenario information.
}
\examples{
produce_marine_nutrient_pm(model_results, "Max Habitat")

}
