% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_access_and_economic_pms.R
\name{produce_percent_harvestable_abv_threshold_pm}
\alias{produce_percent_harvestable_abv_threshold_pm}
\title{Produce Percent Harvestable Above Threshold Performance Metrics}
\usage{
produce_percent_harvestable_abv_threshold_pm(model_results_df, scenario)
}
\arguments{
\item{model_results_df}{A data frame containing the model results.}

\item{scenario}{The scenario name for which the percentage of harvestable years is calculated.}
}
\value{
The percentage of harvestable years above the threshold performance metric and the maximum number of consecutive harvestable years.
}
\description{
Calculate the percentage of harvestable years above a threshold performance metric based on model results.
}
\examples{
model_results <- read.csv("model_results.csv")
percent_harvestable <- produce_percent_harvestable_abv_threshold_pm(model_results, "Scenario1")
print(percent_harvestable)

}
