% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_biological_pms.R
\name{produce_phos_pm}
\alias{produce_phos_pm}
\title{Calculate PHOS Levels}
\usage{
produce_phos_pm(model_results_df)
}
\arguments{
\item{model_results_df}{A data frame containing model results.}
}
\value{
A data frame with the calculated average, minimum, and maximum annual PHOS levels.
}
\description{
This function calculates the average, minimum, and maximum annual Proportion hatchery on spawning grounds (PHOS) levels based on model results.
}
\examples{
produce_phos_pm(model_results)

}
