% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_biological_pms.R
\name{produce_populations_present_pm}
\alias{produce_populations_present_pm}
\title{This function calculates if populations are present on a trib based on model results.}
\usage{
produce_populations_present_pm(model_results_df)
}
\arguments{
\item{model_results_df}{A data frame containing model results.}
}
\value{
A data frame with the calculated populations per scenario.
}
\description{
This function calculates if populations are present on a trib based on model results.
}
\examples{
produce_populations_present_pm(model_results)

}
