% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_biological_pms.R
\name{produce_shannon_div_ind_size_and_timing_pm}
\alias{produce_shannon_div_ind_size_and_timing_pm}
\title{Calculate Shannon Diversity Index for Juvenile Size and Timing}
\usage{
produce_shannon_div_ind_size_and_timing_pm(model_results_df)
}
\arguments{
\item{model_results_df}{A data frame containing model results.}
}
\value{
A data frame with the calculated Shannon Diversity Index (DI) for juvenile size and timing.
}
\description{
This function calculates the Shannon Diversity Index (DI) for juvenile size and timing based on model results.
}
\examples{
produce_shannon_div_ind_size_and_timing_pm(model_results)

}
