% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_access_and_economic_pms.R
\name{produce_spawner_abundance_above_biological_objective_pm}
\alias{produce_spawner_abundance_above_biological_objective_pm}
\title{Produce Spawner Abundance Above Biological Objective Performance Metrics}
\usage{
produce_spawner_abundance_above_biological_objective_pm(
  model_results_df,
  scenario,
  ocean_harvest_seperate = FALSE,
  ocean_harvest
)
}
\arguments{
\item{model_results_df}{A data frame containing the model results.}

\item{scenario}{The scenario name for which the spawner abundance is calculated.}

\item{ocean_harvest_seperate}{Boolean indicating whether ocean harvest is separated from remaining spawners.}

\item{ocean_harvest}{The value of ocean harvest if ocean_harvest_seperate is TRUE.}
}
\value{
A data frame containing the spawner abundance above biological objective performance metrics.
\itemize{
\item \code{avg_remaining_annual_spawners}: The average remaining annual spawners above the biological objective.
\item \code{min_remaining_spawners}: The minimum remaining spawners above the biological objective.
\item \code{max_remaining_spawners}: The maximum remaining spawners above the biological objective.
}
}
\description{
Calculate the spawner abundance above the biological objective performance metrics based on model results.
}
\examples{
model_results <- read.csv("model_results.csv")
spawner_metrics <- produce_spawner_abundance_above_biological_objective_pm(model_results, "Scenario1", TRUE, 100)
print(spawner_metrics)

}
