% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_biological_pms.R
\name{produce_spawner_abundance_pm}
\alias{produce_spawner_abundance_pm}
\title{Calculate Annual Spawner Abundance}
\usage{
produce_spawner_abundance_pm(model_results_df)
}
\arguments{
\item{model_results_df}{A data frame containing model results.}
}
\value{
A data frame with the calculated average, minimum, and maximum annual spawner abundance per scenario.
}
\description{
This function calculates the average, minimum, and maximum annual spawner abundance based on model results.
}
\examples{
produce_spawner_abundance_pm(model_results)

}
