% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_habitat_and_ecological_process_pms.R
\name{produce_spawning_wetted_acre_day_pm}
\alias{produce_spawning_wetted_acre_day_pm}
\title{Calculate Spawning Wetted Acre-Day}
\usage{
produce_spawning_wetted_acre_day_pm(model_parameters, scenario, selected_run)
}
\arguments{
\item{model_parameters}{A list containing the model parameters.}

\item{scenario}{The scenario name for which the spawning wetted acre-day is calculated.}

\item{selected_run}{The selected run for which the spawning wetted acre-day is calculated.}
}
\value{
A data frame containing the spawning wetted acre-day.
}
\description{
Calculate the spawning wetted acre-day based on model parameters, scenario, and selected run.
}
\examples{
model_params <- fallRunDSM::r_to_r_baseline_params
spawn_hab <- produce_spawning_wetted_acre_day_pm(model_params, "Baseline", "fall")
print(spawn_hab)

}
