% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_biological_pms.R
\name{produce_time_to_recovery_pm}
\alias{produce_time_to_recovery_pm}
\title{Calculate Time to Recovery for Two Diversity Groups}
\usage{
produce_time_to_recovery_pm(model_results_df, selected_run)
}
\arguments{
\item{model_results_df}{The model results dataframe.}

\item{selected_run}{The selected run.}
}
\value{
The time to recovery for two diversity groups, or a message indicating that no two diversity groups showed recovery throughout the simulation.
}
\description{
This function calculates the time to recovery for two diversity groups based on the provided model results dataframe and selected run.
}
\examples{
model_results <- data.frame(
  performance_metric = c("Natural Spawners", "Growth Rate Natural Spawners", "PHOS", "CRR: Total Adult to Returning Natural Adult"),
  value = c(100, 1.5, 0.03, 0.8),
  location = c("Location A", "Location B", "Location C", "Location D"),
  year = c(2021, 2022, 2023, 2024),
  scenario = c("Scenario 1", "Scenario 2", "Scenario 1", "Scenario 2"),
  run = c(1, 1, 2, 2)
)

produce_time_to_recovery_pm(model_results, 2)
}
