% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_habitat_and_ecological_process_pms.R
\name{produce_wetted_acre_day_floodplain_activation_pm}
\alias{produce_wetted_acre_day_floodplain_activation_pm}
\title{Produce Wetted Acre-Day Floodplain Activation Performance Metrics}
\usage{
produce_wetted_acre_day_floodplain_activation_pm(
  model_parameters,
  scenario,
  selected_run
)

produce_wetted_acre_day_floodplain_activation_pm(
  model_parameters,
  scenario,
  selected_run
)
}
\arguments{
\item{model_parameters}{A data frame containing the model parameters.}

\item{scenario}{The scenario name for which the wetted acre-day floodplain activation metrics are calculated.}

\item{selected_run}{The selected run for which the wetted acre-day floodplain activation metrics are calculated.}
}
\value{
A data frame containing the wetted acre-day floodplain activation metrics.

A data frame containing the wetted acre-day floodplain activation metrics.
}
\description{
Calculate the wetted acre-day floodplain based on model parameters and scenario.

Calculate the wetted acre-day floodplain activation above 2 - year 30 day flood performance metrics based on model parameters and scenario.
}
\examples{
model_params <- read.csv("model_parameters.csv")
floodplain_activation <- produce_wetted_acre_day_floodplain_activation_pm(model_params, "Scenario1", "Fall")
print(floodplain_activation)

model_params <- read.csv("model_parameters.csv")
floodplain_activation <- produce_wetted_acre_day_floodplain_activation_pm(model_params, "Scenario1", "Fall")
print(floodplain_activation)

}
