% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_habitat_and_ecological_process_pms.R
\name{total_habitat_decay}
\alias{total_habitat_decay}
\title{Calculate Total Habitat Decay}
\usage{
total_habitat_decay(model_params, scenario, run)
}
\arguments{
\item{scenario}{The scenario name for which the total habitat decay is calculated.}

\item{model_parameters}{A list containing the model parameters.}
}
\value{
A data frame containing the total habitat decay.
}
\description{
Calculate the total habitat decay based on model parameters and scenario.
}
\examples{
model_params <- fallRunDSM::r_to_r_baseline_params
total_decay <- total_habitat_decay(model_params, "Baseline")
print(total_decay)

}
