% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process_model_results.R
\name{create_model_results_dataframe}
\alias{create_model_results_dataframe}
\title{Create Model Results Dataframe}
\usage{
create_model_results_dataframe(
  model_results,
  model_parameters,
  scenario_name,
  selected_run
)
}
\arguments{
\item{model_results}{The model results object containing the necessary data.}

\item{model_parameters}{The model parameters object containing the necessary data.}

\item{scenario_name}{The name of the scenario.}

\item{selected_run}{The selected chinook run.}
}
\value{
A dataframe containing performance metrics for different locations and years.
}
\description{
This function takes model results and parameters, along with scenario information, and creates a dataframe containing performance metrics for different locations and years.
}
\examples{
create_model_results_dataframe(model_results, model_parameters, "Scenario 1", "fall")
}
